% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backends.R
\name{getProviders1}
\alias{getProviders1}
\title{CGet Providers
Blah Blah Blah}
\usage{
getProviders1(activeOnly = T, usr = NULL, pwd = NULL)
}
\arguments{
\item{Bob}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
getProviders1
}
\description{
CGet Providers
Blah Blah Blah
}
