% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backends.R
\name{getSoilData}
\alias{getSoilData}
\title{Ctetsysjh}
\usage{
getSoilData(providers = NULL, observedProperty = NULL,
  observedPropertyGroup = NULL, usr = "Admin", pwd = "c")
}
\arguments{
\item{temp_F}{The temperature in Fahrenheit.}
}
\value{
The temperature in Kelvin.
getSoilData
}
\description{
This function converts input temperatures in Fahrenheit to Kelvin.
}
