% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backends.R
\name{getSoilData}
\alias{getSoilData}
\title{Returns Observed Soil Properties}
\usage{
getSoilData(providers = NULL, observedProperty = NULL,
  observedPropertyGroup = NULL, usr = "Public", pwd = "Public")
}
\arguments{
\item{providers}{List of the Providers to query}

\item{observedProperty}{the Observed Soil Property code to query. It can be a ':' delimited list eg 4A1;3B2}

\item{providers}{observedPropertyGroup The Observed Soil property Group to Query on}
}
\value{
Dataframe of Observed Property values
}
\description{
This function will query all the available data providers to return all of the available soil observed property data
}
